factor out context struct

+++ tests/counter.rs
@@ -51,7 +51,7 @@ fn i32_counter() -> TestResult {
     ];
 
     for testcase in testcases {
-        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE as i32)?;
+        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE)?;
         let mut vm = VM::new(&wasm)?;
         vm.set_memory_size(10 * SUBMEMORY_SIZE + wasm_submemory::HEADROOM)?;
         for i in 1..=10 {
@@ -108,7 +108,7 @@ fn f32_counter() -> TestResult {
     ];
 
     for testcase in testcases {
-        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE as i32)?;
+        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE)?;
         let mut vm = VM::new(&wasm)?;
         vm.set_memory_size(10 * SUBMEMORY_SIZE + wasm_submemory::HEADROOM)?;
         for i in 1..=10 {
@@ -151,7 +151,7 @@ fn i64_counter() -> TestResult {
     }];
 
     for testcase in testcases {
-        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE as i32)?;
+        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE)?;
         let mut vm = VM::new(&wasm)?;
         vm.set_memory_size(10 * SUBMEMORY_SIZE + wasm_submemory::HEADROOM)?;
         for i in 1..=10 {
@@ -194,7 +194,7 @@ fn i64_8_counter() -> TestResult {
     }];
 
     for testcase in testcases {
-        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE as i32)?;
+        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE)?;
         let mut vm = VM::new(&wasm)?;
         vm.set_memory_size(10 * SUBMEMORY_SIZE + wasm_submemory::HEADROOM)?;
         for i in 1..=10 {
@@ -237,7 +237,7 @@ fn f64_counter() -> TestResult {
     }];
 
     for testcase in testcases {
-        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE as i32)?;
+        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE)?;
         let mut vm = VM::new(&wasm)?;
         vm.set_memory_size(10 * SUBMEMORY_SIZE + wasm_submemory::HEADROOM)?;
         for i in 1..=10 {
