+++ tests/data.rs
@@ -45,7 +45,7 @@ fn data() -> TestResult {
     ];
 
     for testcase in testcases {
-        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE as i32)?;
+        let wasm = wasm_submemory::rewrite(testcase.wasm, SUBMEMORY_SIZE)?;
         let mut vm = VM::new(&wasm)?;
         vm.set_memory_size(10 * SUBMEMORY_SIZE + wasm_submemory::HEADROOM)?;
         for i in 0..10 {
